requirements.txt
slack-bolt>=1.18.0
discord.py>=2.3.2
httpx>=0.27.0
beautifulsoup4>=4.12.3
anthropic>=0.40.0
apscheduler>=3.10.4
pyyaml>=6.0.2
python-dotenv>=1.0.1
feedparser>=6.0.11
tldextract>=5.1.2
cachetools>=5.5.0
lxml>=5.3.0
